Updating: -Q04 Weather

+++ L10 Regex/L10 Regex Lab (II)/L10 Regex (II)/Q04 Weather/Town.cs
@@ -1,6 +1,6 @@
 ﻿public class Town
 {
     public string Name { get; set; }
-    public float Tempreture { get; set; }
+    public double Tempreture { get; set; }
     public string Forecast { get; set; } 
 }
